telegram-bot/requirements.txt
google-generativeai>=0.8.0
python-telegram-bot>=21.0
httpx>=0.27.0
python-dotenv>=1.0.0
